No sense to have a model delete itself it doesn't exist in the first place.

+++ trunk/www/go/base/data/Store.php
@@ -156,7 +156,8 @@ class GO_Base_Data_Store extends GO_Base_Data_AbstractStore {
           }
 
           $model = $staticModel->findByPk($modelPk);
-          $model->delete();
+					if (!empty($model))
+						$model->delete();
         }
         $this->_response['deleteSuccess'] = true;
       } catch (Exception $e) {
